Add 動的な変更を可能に
