+++ front/src/features/diaries/components/edit-diary-button.tsx
@@ -1,15 +1,27 @@
 import { Button } from '@/components/ui/button'
+import type { UpdateDiaryInput } from '@/features/diaries/api/update-diary'
 import { Ban, Check } from 'lucide-react'
 import type { Dispatch, SetStateAction } from 'react'
+import type { SubmitHandler, UseFormReturn } from 'react-hook-form'
 
 type EditDiaryButtonProps = {
   editFlag: boolean
   setEditFlag: Dispatch<SetStateAction<boolean>>
+  form: UseFormReturn<
+    {
+      body: string
+    },
+    any,
+    undefined
+  >
+  onSubmit: SubmitHandler<UpdateDiaryInput>
 }
 
 export const EditDiaryButton = ({
   editFlag,
   setEditFlag,
+  form,
+  onSubmit
 }: EditDiaryButtonProps) => {
   return (
     <div>
@@ -18,6 +30,7 @@ export const EditDiaryButton = ({
           <Button
             onClick={() => {
               setEditFlag(false)
+              form.handleSubmit(onSubmit)
             }}
             icon={<Check className="sizew-4" />}
           >
